Improve dependencies flow and testing
